+++ app/src/main/java/com/darkliself/engenioustask/repository/remote/UsersGithubRemoteRepository.kt
@@ -1,4 +1,4 @@
-package com.darkliself.engenioustask.repository
+package com.darkliself.engenioustask.repository.remote
 
 import com.darkliself.engenioustask.data.retrofit.api.UserApiService
 import javax.inject.Inject
