+++ app/src/main/java/com/darkliself/engenioustask/repository/remote/UsersRemoteRepository.kt
@@ -1,4 +1,4 @@
-package com.darkliself.engenioustask.repository
+package com.darkliself.engenioustask.repository.remote
 
 import com.darkliself.engenioustask.data.retrofit.model.User
